backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
alembic==1.13.0
numpy==1.24.3
scipy==1.11.4
pandas==2.1.4
python-dotenv==1.0.0